requirements.txt
asgiref==3.8.1
cffi==1.17.1
cloudinary==1.41.0
cryptography==43.0.1
dj-database-url==0.5.0
dj-rest-auth==5.0.1
Django==4.2
django-allauth==0.55.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==2.0.0
setuptools==75.1.0
sqlparse==0.5.1
urllib3==2.2.3
